the restructuring file structure done
