+++ src/assets/PeopleList.js
