+++ src/store/reducers/peopleReducers.js
